Refractor rentals code
