+++ modules/creator.rb
@@ -40,6 +40,30 @@ module Creator
     Teacher.new(age, spec, name)
   end
 
+# private methods used for rental
+
+  def select_book
+    puts 'Select a book from the list by [number]:'
+    @books.each_with_index { |book, i| puts "[#{i}] #{book.title} by #{book.author}" }
+    print 'Enter [number]: '
+    index = gets.chomp.to_i
+    return @books[index] if @books[index]
+
+    select_book
+  end
+
+  def select_person
+    puts 'Select a person from the list by [number]:'
+    @persons.each_with_index do |person, i|
+      puts "[#{i}] [#{person.class.name}] ID: #{person.id} Name: #{person.name}"
+    end
+    print 'Enter [number]: '
+    index = gets.chomp.to_i
+    return @persons[index] if @persons[index]
+
+    select_person
+  end
+
   public
 
   def create_new_person
@@ -63,4 +87,7 @@ module Creator
     author = gets.chomp
     Book.new(title, author)
   end
+
+  # create rentals
+  
 end
